+++ resqpy/organize/__init__.py
@@ -2,23 +2,23 @@ __all__ = [
     'OrganizationFeature',
 ]
 
-from .OrganizationFeature import OrganizationFeature
-from .GeobodyFeature import GeobodyFeature
-from .BoundaryFeature import BoundaryFeature
+from .organization_feature import OrganizationFeature
+from .geobody_feature import GeobodyFeature
+from .boundary_feature import BoundaryFeature
 
-from .FrontierFeature import FrontierFeature
-from .GeologicUnitFeature import GeologicUnitFeature
-from .FluidBoundaryFeature import FluidBoundaryFeature
-from .RockFluidUnitFeature import RockFluidUnitFeature
-from .TectonicBoundaryFeature import TectonicBoundaryFeature
-from .GeneticBoundaryFeature import GeneticBoundaryFeature
-from .WellboreFeature import WellboreFeature
-from .FaultInterpretation import FaultInterpretation
-from .EarthModelInterpretation import EarthModelInterpretation
-from .HorizonInterpretation import HorizonInterpretation
-from .GeobodyBoundaryInterpretation import GeobodyBoundaryInterpretation
-from .GeobodyInterpretation import GeobodyInterpretation
-from .WellboreInterpretation import WellboreInterpretation
+from .frontier_feature import FrontierFeature
+from .geologic_unit_feature import GeologicUnitFeature
+from .fluid_boundary_feature import FluidBoundaryFeature
+from .rock_fluid_unit_feature import RockFluidUnitFeature
+from .tectonic_boundary_feature import TectonicBoundaryFeature
+from .genetic_boundary_feature import GeneticBoundaryFeature
+from .wellbore_feature import WellboreFeature
+from .fault_interpretation import FaultInterpretation
+from .earth_model_interpretation import EarthModelInterpretation
+from .horizon_interpretation import HorizonInterpretation
+from .geobody_boundary_interpretation import GeobodyBoundaryInterpretation
+from .geobody_interpretation import GeobodyInterpretation
+from .wellbore_interpretation import WellboreInterpretation
 
 from resqpy.organize.organize_functions import (equivalent_extra_metadata, alias_for_attribute,
                                                 extract_has_occurred_during, equivalent_chrono_pairs,
