+++ resqpy/organize/fault_interpretation.py
@@ -7,7 +7,7 @@ import resqpy.olio.uuid as bu
 import resqpy.olio.xml_et as rqet
 from resqpy.olio.base import BaseResqpy
 from resqpy.olio.xml_namespaces import curly_namespace as ns
-from .TectonicBoundaryFeature import TectonicBoundaryFeature
+from .tectonic_boundary_feature import TectonicBoundaryFeature
 
 
 class FaultInterpretation(BaseResqpy):
