+++ resqpy/organize/rock_fluid_unit_feature.py
@@ -5,7 +5,7 @@ import resqpy.olio.uuid as bu
 import resqpy.olio.xml_et as rqet
 from resqpy.olio.base import BaseResqpy
 from resqpy.olio.xml_namespaces import curly_namespace as ns
-from .BoundaryFeature import BoundaryFeature
+from .boundary_feature import BoundaryFeature
 
 
 class RockFluidUnitFeature(BaseResqpy):
